+++ packages/hot-updater/src/commands/doctor.ts
@@ -1,5 +1,4 @@
-import * as p from "@clack/prompts";
-import { getCwd } from "@hot-updater/plugin-core";
+import { getCwd, p } from "@hot-updater/cli-tools";
 import { merge } from "es-toolkit";
 import fs from "fs";
 import { readPackageUp } from "read-package-up";