+++ packages/hot-updater/src/config.ts
@@ -1,7 +1,5 @@
-import type {
-  ConfigInput,
-  HotUpdaterConfigOptions,
-} from "@hot-updater/plugin-core";
+import type { HotUpdaterConfigOptions } from "@hot-updater/cli-tools";
+import type { ConfigInput } from "@hot-updater/plugin-core";
 
 export const defineConfig = (
   config: ConfigInput | ((options: HotUpdaterConfigOptions) => ConfigInput),