+++ plugins/repack/src/index.ts
@@ -1,7 +1,7 @@
+import { colors } from "@hot-updater/cli-tools";
 import type { Compiler, RspackPluginInstance } from "@rspack/core";
 import fs from "fs";
 import path from "path";
-import picocolors from "picocolors";
 import { uuidv7 } from "uuidv7";
 
 const NIL_UUID = "00000000-0000-0000-0000-000000000000";
@@ -20,9 +20,7 @@ const getBundleId = () => {
     bundleId = fs.readFileSync(bundleIdPath, "utf-8");
   } else {
     fs.writeFileSync(bundleIdPath, bundleId);
-    console.log(
-      picocolors.green(`[HotUpdater] Generated bundle ID: ${bundleId}`),
-    );
+    console.log(colors.green(`[HotUpdater] Generated bundle ID: ${bundleId}`));
   }
 
   return bundleId;