+++ examples/v0.81.0/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @hot-updater/example-react-native-v81
 
+## 0.21.8
+
+### Patch Changes
+
+- @hot-updater/react-native@0.21.8
+
+## 0.21.7
+
+### Patch Changes
+
+- Updated dependencies [2b408f2]
+  - @hot-updater/react-native@0.21.7
+
+## 0.21.6
+
+### Patch Changes
+
+- 3e9681c: fix(android): Android API 25 compatibility with TarStream
+- Updated dependencies [3e9681c]
+  - @hot-updater/react-native@0.21.6
+
 ## 0.21.5
 
 ### Patch Changes