+++ packages/hot-updater/CHANGELOG.md
@@ -1,5 +1,46 @@
 # hot-updater
 
+## 0.21.8
+
+### Patch Changes
+
+- 3fe8c81: feat(plugin-core): reduced deps for edge-runtime
+- Updated dependencies [3fe8c81]
+  - @hot-updater/plugin-core@0.21.8
+  - @hot-updater/cli-tools@0.21.8
+  - @hot-updater/cloudflare@0.21.8
+  - @hot-updater/firebase@0.21.8
+  - @hot-updater/aws@0.21.8
+  - @hot-updater/console@0.21.8
+  - @hot-updater/supabase@0.21.8
+  - @hot-updater/core@0.21.8
+
+## 0.21.7
+
+### Patch Changes
+
+- Updated dependencies [2b408f2]
+  - @hot-updater/plugin-core@0.21.7
+  - @hot-updater/cloudflare@0.21.7
+  - @hot-updater/firebase@0.21.7
+  - @hot-updater/supabase@0.21.7
+  - @hot-updater/aws@0.21.7
+  - @hot-updater/console@0.21.7
+  - @hot-updater/core@0.21.7
+
+## 0.21.6
+
+### Patch Changes
+
+- b12394d: feat(cli): create migration sql hot-updater generate-db
+  - @hot-updater/console@0.21.6
+  - @hot-updater/core@0.21.6
+  - @hot-updater/aws@0.21.6
+  - @hot-updater/cloudflare@0.21.6
+  - @hot-updater/firebase@0.21.6
+  - @hot-updater/plugin-core@0.21.6
+  - @hot-updater/supabase@0.21.6
+
 ## 0.21.5
 
 ### Patch Changes