+++ packages/server/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @hot-updater/server
 
+## 0.21.8
+
+### Patch Changes
+
+- Updated dependencies [3fe8c81]
+  - @hot-updater/plugin-core@0.21.8
+  - @hot-updater/core@0.21.8
+
+## 0.21.7
+
+### Patch Changes
+
+- 2b408f2: docs: revamp hot-updater.dev
+- Updated dependencies [2b408f2]
+  - @hot-updater/plugin-core@0.21.7
+  - @hot-updater/core@0.21.7
+
+## 0.21.6
+
+### Patch Changes
+
+- b12394d: feat(cli): create migration sql hot-updater generate-db
+- d4c23bc: fix(server): id column uuid
+  - @hot-updater/core@0.21.6
+  - @hot-updater/plugin-core@0.21.6
+
 ## 0.21.5
 
 ### Patch Changes