+++ plugins/aws/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @hot-updater/aws
 
+## 0.21.8
+
+### Patch Changes
+
+- 3fe8c81: feat(plugin-core): reduced deps for edge-runtime
+- Updated dependencies [3fe8c81]
+  - @hot-updater/plugin-core@0.21.8
+  - @hot-updater/cli-tools@0.21.8
+
+## 0.21.7
+
+### Patch Changes
+
+- 2b408f2: docs: revamp hot-updater.dev
+- Updated dependencies [2b408f2]
+  - @hot-updater/plugin-core@0.21.7
+
+## 0.21.6
+
+### Patch Changes
+
+- @hot-updater/plugin-core@0.21.6
+
 ## 0.21.5
 
 ### Patch Changes